revert(modelopt): remove review-driven scope
