+++ tests/unit/models/generation/test_vllm_modelopt_real_quant_config.py
@@ -2391,243 +2391,6 @@ def test_resolve_quant_cfg_resolves_repo_relative_recipe_from_remote_cwd(
     assert captured["config_file"] == str((repo_root / quant_cfg).resolve())
 
 
-def test_resolve_quant_cfg_probes_suffix_for_repo_relative_recipe(
-    monkeypatch, tmp_path
-):
-    modelopt_recipe = pytest.importorskip("modelopt.recipe")
-    captured = {}
-
-    def fake_load_config(config_file):
-        captured["config_file"] = config_file
-        return {"quant_cfg": [{"name": "mock"}], "algorithm": "max"}
-
-    monkeypatch.setattr(modelopt_recipe, "load_config", fake_load_config)
-    monkeypatch.chdir(tmp_path)
-
-    quant_cfg = "examples/modelopt/quant_configs/nvfp4_a16"
-    resolve_quant_cfg(quant_cfg)
-
-    repo_root = Path(modelopt_utils.__file__).resolve().parents[2]
-    assert captured["config_file"] == str((repo_root / f"{quant_cfg}.yaml").resolve())
-
-
-def test_resolve_quant_cfg_prefers_current_cwd_relative_recipe(monkeypatch, tmp_path):
-    modelopt_recipe = pytest.importorskip("modelopt.recipe")
-    captured = {}
-
-    def fake_load_config(config_file):
-        captured["config_file"] = config_file
-        return {"quant_cfg": [{"name": "mock"}], "algorithm": "max"}
-
-    monkeypatch.setattr(modelopt_recipe, "load_config", fake_load_config)
-    quant_cfg = Path("examples/modelopt/quant_configs/nvfp4_a16.yaml")
-    cwd_recipe = tmp_path / quant_cfg
-    cwd_recipe.parent.mkdir(parents=True)
-    cwd_recipe.touch()
-    monkeypatch.chdir(tmp_path)
-
-    resolve_quant_cfg(str(quant_cfg))
-
-    assert captured["config_file"] == str(quant_cfg)
-
-
-def test_resolve_quant_cfg_preserves_current_cwd_suffix_probe(monkeypatch, tmp_path):
-    modelopt_recipe = pytest.importorskip("modelopt.recipe")
-    captured = {}
-
-    def fake_load_config(config_file):
-        captured["config_file"] = config_file
-        return {"quant_cfg": [{"name": "mock"}], "algorithm": "max"}
-
-    monkeypatch.setattr(modelopt_recipe, "load_config", fake_load_config)
-    quant_cfg = Path("custom/recipe")
-    cwd_recipe = tmp_path / f"{quant_cfg}.yaml"
-    cwd_recipe.parent.mkdir(parents=True)
-    cwd_recipe.touch()
-    monkeypatch.chdir(tmp_path)
-
-    resolve_quant_cfg(str(quant_cfg))
-
-    assert captured["config_file"] == str(quant_cfg)
-
-
-def test_resolve_quant_cfg_ignores_exact_extensionless_repo_file(monkeypatch, tmp_path):
-    modelopt_recipe = pytest.importorskip("modelopt.recipe")
-    captured = {}
-
-    def fake_load_config(config_file):
-        captured["config_file"] = config_file
-        return {"quant_cfg": [{"name": "mock"}], "algorithm": "max"}
-
-    monkeypatch.setattr(modelopt_recipe, "load_config", fake_load_config)
-    repo_module = tmp_path / "repo/nemo_rl/modelopt/utils.py"
-    repo_module.parent.mkdir(parents=True)
-    repo_module.touch()
-    monkeypatch.setattr(modelopt_utils, "__file__", str(repo_module))
-    quant_cfg = Path("recipes/custom")
-    exact_recipe = tmp_path / "repo" / quant_cfg
-    exact_recipe.parent.mkdir(parents=True)
-    exact_recipe.touch()
-    yaml_recipe = tmp_path / "repo" / f"{quant_cfg}.yaml"
-    yaml_recipe.touch()
-    remote_cwd = tmp_path / "workers/remote"
-    remote_cwd.mkdir(parents=True)
-    monkeypatch.chdir(remote_cwd)
-
-    resolve_quant_cfg(str(quant_cfg))
-
-    assert captured["config_file"] == str(yaml_recipe.resolve())
-
-
-def test_resolve_quant_cfg_cache_suffix_hashes_repo_relative_recipe(
-    tmp_path, monkeypatch
-):
-    modelopt_recipe = pytest.importorskip("modelopt.recipe")
-    repo_module = tmp_path / "repo/nemo_rl/modelopt/utils.py"
-    repo_module.parent.mkdir(parents=True)
-    repo_module.touch()
-    monkeypatch.setattr(modelopt_utils, "__file__", str(repo_module))
-    recipe = tmp_path / "repo/recipes/custom.yaml"
-    recipe.parent.mkdir(parents=True)
-    remote_cwd = tmp_path / "workers/remote"
-    remote_cwd.mkdir(parents=True)
-    monkeypatch.chdir(remote_cwd)
-    config = {"quant_cfg": "recipes/custom.yaml"}
-
-    def fake_load_config(config_file):
-        algorithm = (
-            "smoothquant" if "smoothquant" in Path(config_file).read_text() else "max"
-        )
-        return {"quant_cfg": [{"name": "mock"}], "algorithm": algorithm}
-
-    monkeypatch.setattr(modelopt_recipe, "load_config", fake_load_config)
-    recipe.write_text("quantize:\n  algorithm: max\n")
-    first_suffix = modelopt_utils._quant_checkpoint_cache_suffix(config)
-    recipe.write_text("quantize:\n  algorithm: smoothquant\n")
-
-    assert modelopt_utils._quant_checkpoint_cache_suffix(config) != first_suffix
-
-
-def test_resolve_quant_cfg_cache_suffix_hashes_imported_semantics(
-    tmp_path, monkeypatch
-):
-    modelopt_recipe = pytest.importorskip("modelopt.recipe")
-    recipe = tmp_path / "recipe.yaml"
-    dependency = tmp_path / "dependency.yaml"
-    recipe.write_text("imports:\n  dependency: dependency.yaml\n")
-
-    def fake_load_config(_config_file):
-        algorithm = dependency.read_text().strip()
-        return {"quant_cfg": [{"name": "mock"}], "algorithm": algorithm}
-
-    monkeypatch.setattr(modelopt_recipe, "load_config", fake_load_config)
-    dependency.write_text("max")
-    config = {"quant_cfg": str(recipe)}
-    first_resolved = resolve_quant_cfg(str(recipe))
-    first_suffix = modelopt_utils._quant_checkpoint_cache_suffix(config)
-    dependency.write_text("smoothquant")
-
-    assert resolve_quant_cfg(str(recipe)) != first_resolved
-    assert modelopt_utils._quant_checkpoint_cache_suffix(config) != first_suffix
-
-
-def test_resolve_quant_cfg_cache_suffix_ignores_builtin_name_file_collision(
-    tmp_path, monkeypatch
-):
-    mtq = sys.modules["modelopt.torch.quantization"]
-
-    monkeypatch.setattr(
-        mtq,
-        "TEST_BUILTIN_CFG",
-        {"quant_cfg": [{"name": "builtin"}], "algorithm": "max"},
-        raising=False,
-    )
-    collision = tmp_path / "TEST_BUILTIN_CFG.yaml"
-    monkeypatch.chdir(tmp_path)
-    config = {"quant_cfg": "TEST_BUILTIN_CFG"}
-    collision.write_text("unrelated-v1")
-    first_suffix = modelopt_utils._quant_checkpoint_cache_suffix(config)
-    collision.write_text("unrelated-v2")
-
-    assert resolve_quant_cfg("TEST_BUILTIN_CFG") == {
-        "quant_cfg": [{"name": "builtin"}],
-        "algorithm": "max",
-    }
-    assert modelopt_utils._quant_checkpoint_cache_suffix(config) == first_suffix
-
-
-def test_resolve_quant_cfg_cache_suffix_handles_mixed_mapping_keys(monkeypatch):
-    mtq = sys.modules["modelopt.torch.quantization"]
-    monkeypatch.setattr(
-        mtq,
-        "TEST_MIXED_KEYS",
-        {
-            "algorithm": "max",
-            "quant_cfg": [{"cfg": {"block_sizes": {-1: 16, "type": "dynamic"}}}],
-        },
-        raising=False,
-    )
-
-    suffix = modelopt_utils._quant_checkpoint_cache_suffix(
-        {"quant_cfg": "TEST_MIXED_KEYS"}
-    )
-
-    assert suffix.startswith("_modelopt_")
-
-
-def test_resolve_quant_cfg_cache_suffix_preserves_legacy_quant_cfg_order(monkeypatch):
-    mtq = sys.modules["modelopt.torch.quantization"]
-    resolved = {
-        "algorithm": "max",
-        "quant_cfg": {
-            "*": {"enable": False},
-            "*weight_quantizer": {"enable": True},
-        },
-    }
-    monkeypatch.setattr(mtq, "TEST_LEGACY_ORDER", resolved, raising=False)
-
-    first_suffix = modelopt_utils._quant_checkpoint_cache_suffix(
-        {"quant_cfg": "TEST_LEGACY_ORDER"}
-    )
-    resolved["quant_cfg"] = {
-        "*weight_quantizer": {"enable": True},
-        "*": {"enable": False},
-    }
-
-    assert (
-        modelopt_utils._quant_checkpoint_cache_suffix(
-            {"quant_cfg": "TEST_LEGACY_ORDER"}
-        )
-        != first_suffix
-    )
-
-
-def test_resolve_quant_cfg_does_not_resolve_path_outside_checkout(
-    monkeypatch, tmp_path
-):
-    modelopt_recipe = pytest.importorskip("modelopt.recipe")
-    captured = {}
-
-    def fake_load_config(config_file):
-        captured["config_file"] = config_file
-        return {"quant_cfg": [{"name": "mock"}], "algorithm": "max"}
-
-    monkeypatch.setattr(modelopt_recipe, "load_config", fake_load_config)
-    repo_module = tmp_path / "repo/nemo_rl/modelopt/utils.py"
-    repo_module.parent.mkdir(parents=True)
-    repo_module.touch()
-    monkeypatch.setattr(modelopt_utils, "__file__", str(repo_module))
-    (tmp_path / "outside.yaml").touch()
-    remote_cwd = tmp_path / "workers/remote"
-    remote_cwd.mkdir(parents=True)
-    monkeypatch.chdir(remote_cwd)
-
-    quant_cfg = "../outside.yaml"
-    resolve_quant_cfg(quant_cfg)
-
-    assert captured["config_file"] == quant_cfg
-
-
 def test_resolve_quant_cfg_accepts_builtin_modelopt_constant(monkeypatch):
     mtq = pytest.importorskip("modelopt.torch.quantization")
     sentinel = {"quant_cfg": [{"name": "builtin"}], "algorithm": "max"}
@@ -2679,10 +2442,7 @@ def test_resolve_quant_cfg_rejects_unknown_config(monkeypatch):
 
     monkeypatch.setattr(modelopt_recipe, "load_config", fake_load_config)
 
-    with pytest.raises(
-        ValueError,
-        match="absolute, relative to the current working directory, or relative to the NeMo-RL checkout",
-    ):
+    with pytest.raises(ValueError, match="Unknown quant_cfg"):
         resolve_quant_cfg("does-not-exist")
